requirements.txt
grpcio==1.37.1
grpcio-tools==1.37.1
mmh3==3.0.0
mongoengine==0.24.1
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
protobuf==3.20.1
pymilvus==2.0.2
pymongo==4.1.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
ujson==5.1.0
